List books and extract books logic to a module
