+++ start.rb
@@ -8,13 +8,13 @@ class Start
   def start_app(selection)
     case selection
     when 1
-      @app.list_all_books
+      @app.books.list_all_books
     when 2
       @app.list_all_people
     when 3
       @app.create_person
     when 4
-      @app.create_book
+      @app.books.create_book
     when 5
       @app.create_rental
     when 6
