+++ storage/books.json
@@ -1,1 +1,1 @@
-[{ "json_class": "Book", "title": "aaaaa", "author": "bbbb" }]
+[{"json_class":"Book","title":"aaaaa","author":"bbbb"},{"json_class":"Book","title":"River ber","author":"Alpha"}]
